// crm-frontend/lib/types/dashboard.ts
export interface DbUser {
  id: string;
  close_user_id: string;
  first_name: string;
  last_name: string;
  email: string;
  role: "caller" | "setter" | "closer" | "admin";
  status: "active" | "inactive";
}